% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_fusion.R
\name{bigeigen}
\alias{bigeigen}
\title{Big eigenvalue calculation}
\usage{
bigeigen(X, method = "RSpectra")
}
\arguments{
\item{X}{matrix to be evaluated (can be a Matrix object).}

\item{method}{One of 'irlba' or 'RSpectra'}
}
\value{
The maximal eigenvalue.
}
\description{
Calculate maximal eigenvalue of \code{t(X) \%*\% X} for big
matrices using singular value decomposition.
}
