% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_fusion.R
\name{extract.lme.cov}
\alias{extract.lme.cov}
\title{Function for extracting Z, G, and V from an lme object.}
\usage{
extract.lme.cov(b, data)
}
\arguments{
\item{b}{A fitted object of class 'lme'.}

\item{data}{The data that we want to calculate Z and V.}
}
\value{
A list with elements V, Z and G.
}
\description{
This function has been adapted from the same function in
the package mgcv, to allow for extraction of more parameters. It calculates
the variance-covariance matrix Z of the data, and also returns the
design matrix of random effects Z and the covariance matrix of the random
effects G.
}
