% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-cluster.R, R/methods-bioconductor.R
\name{fuse.cluster}
\alias{fuse.cluster}
\alias{fuse.cluster.default}
\alias{fuse.cluster.BSseq}
\alias{fuse.cluster.methrix}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
fuse.cluster(x, ...)

\method{fuse.cluster}{default}(x, K1, chr = NULL, pos = NULL, ...)

\method{fuse.cluster}{BSseq}(x, ...)

\method{fuse.cluster}{methrix}(x, ...)
}
\arguments{
\item{x}{Input object. One of:
\describe{
\item{matrix}{Unmethylated count matrix (K0).}
\item{BSseq}{A \code{BSseq} object.}
\item{methrix}{A \code{methrix} object.}
}}

\item{...}{K1 Integer or numeric matrix with methylated counts
#' @param ... Additional arguments depending on input type:
\describe{
\item{K1}{Methylated count matrix (required if \code{x} is a matrix).}
\item{chr}{(Optinal) Chromosome labels, one per CpG (matrix input only).}
\item{pos}{(Optional) positions, one per CpG (matrix input only).}
}}

\item{K1}{Methylated count matrix}

\item{chr}{Optional chromosome vector}

\item{pos}{Optional position vector}
}
\value{
Clustering tree as a matrix, with the same structure as hclust, including the columns
\describe{
\item{m1}{ID of first data point. <0 is original point, >0 is row of previous merge.}
\item{m2}{ID of second data point. <0 is original point, >0 is row of previous merge.}
\item{logl_tot}{Change in total log-likelihood of merge.}
\item{logl_merge}{Total cost of points in this merge.}
\item{genomic_dist}{Penalty for genomic distance between points.}
}
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
\examples{
# Example: Clustering generated data
set.seed(1234)
K0 <- matrix(
  rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
K1 <- matrix(
  rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)
tree

}
