% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_fusion.R
\name{fusedL2DescentGLMNet}
\alias{fusedL2DescentGLMNet}
\title{Optimise the fused L2 model with glmnet (using transformed input data)}
\usage{
fusedL2DescentGLMNet(transformed.x, transformed.y, num.groups,
  group.names = 1:num.groups, lambda.sparse, ...)
}
\arguments{
\item{num.groups}{Number of groups (K)}

\item{group.names}{Names for the groups (optional)}

\item{lambda.sparse}{Sparsity penalty hyperparameter}

\item{...}{Further options passed to glmnet.}

\item{new.x}{Transformed covariates (output of generateBlockDiagonalMatrices)}

\item{new.y}{Transformed response (output of generateBlockDiagonalMatrices)}
}
\value{
Matrix of fitted beta values.
}
\description{
Optimise the fused L2 model with glmnet (using transformed input data)
}

