% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_fusion.R
\name{fusedLassoProximalIterationsTaken}
\alias{fusedLassoProximalIterationsTaken}
\title{Following a call to fusedLassoProximal, returns the actual number of iterations taken.}
\usage{
fusedLassoProximalIterationsTaken()
}
\value{
Number of iterations performed in the previous call to fusedLassoProximal.
}
\description{
Following a call to fusedLassoProximal, returns the actual number of iterations taken.
}
