% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_fusion.R
\name{fusionLMM}
\alias{fusionLMM}
\title{Fit a linear mixed model (using the nlme package) with
fusion over the fixed effects. Only l2 fusion is currently implemented.}
\usage{
fusionLMM(formula, random, data, groups, gamma, G = matrix(1,
  length(unique(groups)), length(unique(groups))))
}
\arguments{
\item{formula}{A two-sided linear formula object describing the fixed effects,
identical to the input of the 'lme' function. Note that we have assumed a single
response variable.}

\item{random}{A one-sided formula or list of one-sided formulas describing the
random effects. Identical to the input of the 'lme' function.}

\item{data}{A data frame containing the variables names in fixed and random.}

\item{groups}{A vector of group indicators (length n)}

\item{gamma}{The fusion hyperparameter (must be a scalar)}

\item{G}{Matrix of pairwise group information sharing weights (K by K).
Default is equal sharing across all pairs.}
}
\value{
A list of class fused.lmm, with elements 'beta', the fixed effects
estimated with fusion penalty information sharing across groups, and
'lme.fits', the fitted lme models for each group. Note that the lme fits
will also contain estimates of the fixed effects, but estimated without
information sharing.
}
\description{
A linear mixed effects model is fitted for each group using
the lme function of package nlme. Subsequently, the l2 fusion approach is
applied to share information about the fixed effects among the groups.

Unlike the functions for fusion estimation in fixed effect linear models,
this function does not currently scale to high-dimensional data with a
large number of fixed effects. Consequently, l1 penalisation of the
fixed effects has not been implemented because it is assumed that n > p.
}
