% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island.R
\name{island.cluster}
\alias{island.cluster}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
island.cluster(K0, K1)
}
\arguments{
\item{K0}{Integer matrix with unmethylated counts}

\item{K1}{Integer matrix with methylated counts}
}
\value{
Unsorted clustering tree as a matrix
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
