% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island.R
\name{island.cutree}
\alias{island.cutree}
\title{Cut Hierarchical Clustering Tree into Clusters}
\usage{
island.cutree(tree, k)
}
\arguments{
\item{tree}{Sorted clustering tree}

\item{k}{Number of clusters}
}
\value{
A vector indicating which cluster each element in the original dataframe belonged to
}
\description{
Divides the clustering tree into a specified number of clusters.
}
