% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island.R
\name{island.sort.tree}
\alias{island.sort.tree}
\title{Sort Clustering Tree in Optimal Order}
\usage{
island.sort.tree(tree)
}
\arguments{
\item{tree}{Unsorted clustering tree}
}
\value{
Sorted tree
}
\description{
Sorts the clustering tree produced by \code{\link{island.cluster}}.
}
