% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_fusion.R
\name{predictBLUP}
\alias{predictBLUP}
\title{Prediction method for observed individuals}
\usage{
predictBLUP(object, newdata.fixed, newdata.Z, groups, ...)
}
\arguments{
\item{object}{An S3 object of class 'fused.lmm'}

\item{newdata.fixed}{Model matrix for the fixed effects}

\item{groups}{A vector of group indicators (length n). Note that group
indicators must be identical to the ones used to train the fused.lmm
model.}

\item{newdata.random}{Model matrix for the random effects}
}
\value{
A matrix of fixed effects for each group.
}
\description{
BLUP estimation of the response for new covariate data.
}
