% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_fusion.R
\name{refitLMMFixedEffects}
\alias{refitLMMFixedEffects}
\title{Apply fusion to fixed effects of a linear mixed model.}
\usage{
refitLMMFixedEffects(X, Y, V, groups, gamma = 0, G, eps = 1e-05)
}
\arguments{
\item{X}{Fixed effect covariates (rhs of the formula)}

\item{Y}{Response variable}

\item{V}{variance-covariance matrices of the response, a list of K n_k
by n_k matrices, where K is the number of groups.}

\item{groups}{A vector of group indicators (length n)}

\item{gamma}{The fusion hyperparameter (must be a scalar)}

\item{G}{Matrix of pairwise group information sharing weights (K by K)}

\item{eps}{Convergence threshold for the gradient descent algorithm}
}
\value{
A matrix of fixed effects for each group.
}
\description{
In order to use
this function, we first require standard fitting of a linear mixed model
for each group and then estimation of the variance-covariance matrices
of the response.
}
