% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_fusion.R
\name{bigeigen}
\alias{bigeigen}
\title{Calculate maximal eigenvalue for big matrices using singular
value decomposition}
\usage{
bigeigen(X, method = "RSpectra")
}
\arguments{
\item{X}{Matrix to be evaluated (can be a bigmatrix)}

\item{method}{One of 'irlba' or 'RSpectra'}
}
\value{
The maximal eigenvalue
}
\description{
Calculate maximal eigenvalue for big matrices using singular
value decomposition
}
