% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_fusion.R
\name{generateBlockDiagonalMatrices}
\alias{generateBlockDiagonalMatrices}
\title{Generate block diagonal matrices to allow for fused L2 optimization with glmnet.}
\usage{
generateBlockDiagonalMatrices(X, Y, groups, G, intercept = FALSE,
  penalty.factors = rep(1, dim(X)[2]))
}
\arguments{
\item{X}{covariates matrix (n by p).}

\item{groups}{vector of group indicators (ideally factors, length n)}

\item{G}{matrix representing the fusion strengths between pairs of 
groups (K by K). Zero entries are assumed to be independent pairs.}

\item{intercept}{whether to include an (per-group) intercept in the model}

\item{penalty.factors}{vector of weights for the penalization of 
each covariate (length p)}

\item{response}{vector (length n).}
}
\value{
A list with components X, Y, X.fused and penalty, where
X is a n by pK block-diagonal bigmatrix, Y is a 
re-arranged bigvector of length n, and X.fused is a 
choose(K,2)*p by pK bigmatrix encoding the fusion penalties.
}
\description{
Generate block diagonal matrices to allow for fused L2 optimization with glmnet.
}
\examples{
add(1, 1)
add(10, 1)

}

